+++ reactos/ntoskrnl/kdbg/kdb_serial.c
@@ -1,4 +1,5 @@
-/*
+/* $Id$
+ *
  * COPYRIGHT:       See COPYING in the top level directory
  * PROJECT:         ReactOS kernel
  * FILE:            ntoskrnl/dbg/kdb_serial.c
@@ -16,14 +17,14 @@
 #include <debug.h>
 
 CHAR
-KdbpTryGetCharSerial(ULONG Retry)
+KdbpTryGetCharSerial(UINT Retry)
 {
-    CHAR Result = -1;
+  CHAR Result = -1;
 
-    if (Retry == 0)
-        while (!KdPortGetByteEx(&SerialPortInfo, (PUCHAR)&Result));
-    else
-        while (!KdPortGetByteEx(&SerialPortInfo, (PUCHAR)&Result) && Retry-- > 0);
+  if (Retry == 0)
+     while (!KdPortGetByteEx(&SerialPortInfo, (PUCHAR)&Result));
+  else
+     while (!KdPortGetByteEx(&SerialPortInfo, (PUCHAR)&Result) && Retry-- > 0);
 
-    return Result;
+  return Result;
 }
